+++ prnaseqtools/modes/wgbs.py
@@ -6,6 +6,7 @@ Bismark alignment → methylation extraction → DMRcaller.
 import os
 import sys
 import glob as globmod
+import shutil
 import subprocess
 import math
 from pathlib import Path
@@ -47,6 +48,11 @@ def run(opts):
     if not nomapping:
         tee.write("\nBuilding indices...\n")
         fasta_path = os.path.join(prefix, "reference", f"{genome}_chr_all.fasta")
+        if not os.path.exists(fasta_path):
+            tee.write(f"Error: reference FASTA not found at {fasta_path}\n")
+            return
+        if os.path.islink(f"{genome}.fasta") or os.path.exists(f"{genome}.fasta"):
+            os.unlink(f"{genome}.fasta")
         os.symlink(fasta_path, f"{genome}.fasta")
         run_cmd("bismark_genome_preparation .")
 
@@ -65,13 +71,19 @@ def run(opts):
                         run_cmd(
                             f"cutadapt -j {thread} -m 20 --trim-n -a {adaptor} "
                             f"-o {tag}_trimmed.fastq {tag}.fastq")
-                        os.rename(f"{tag}_trimmed.fastq", f"{tag}.fastq")
+                        if os.path.exists(f"{tag}_trimmed.fastq"):
+                            os.rename(f"{tag}_trimmed.fastq", f"{tag}.fastq")
+                        else:
+                            tee.write(f"Warning: cutadapt did not produce trimmed file for {tag}\n")
 
                     run_cmd(
                         f"bismark -p {thread} -N 1 . {tag}.fastq")
                     run_cmd(
                         f"deduplicate_bismark -s --bam {tag}_bismark_bt2.bam")
-                    os.rename(f"{tag}_bismark_bt2.deduplicated.bam", f"{tag}.bam")
+                    if os.path.exists(f"{tag}_bismark_bt2.deduplicated.bam"):
+                        os.rename(f"{tag}_bismark_bt2.deduplicated.bam", f"{tag}.bam")
+                    else:
+                        tee.write(f"Warning: deduplicate did not produce output for {tag}\n")
                     run_cmd(
                         f"bismark_methylation_extractor --parallel {thread} -s --bedGraph "
                         f"--cutoff 4 --cytosine_report --CX --genome_folder . {tag}.bam")
@@ -86,14 +98,23 @@ def run(opts):
                             f"cutadapt -j {thread} -m 20 --trim-n -a {adaptor} -A {adaptor} "
                             f"-o {tag}_R1_trimmed.fastq -p {tag}_R2_trimmed.fastq "
                             f"{tag}_R1.fastq {tag}_R2.fastq")
-                        os.rename(f"{tag}_R1_trimmed.fastq", f"{tag}_R1.fastq")
-                        os.rename(f"{tag}_R2_trimmed.fastq", f"{tag}_R2.fastq")
+                        if os.path.exists(f"{tag}_R1_trimmed.fastq"):
+                            os.rename(f"{tag}_R1_trimmed.fastq", f"{tag}_R1.fastq")
+                        else:
+                            tee.write(f"Warning: cutadapt did not produce trimmed R1 for {tag}\n")
+                        if os.path.exists(f"{tag}_R2_trimmed.fastq"):
+                            os.rename(f"{tag}_R2_trimmed.fastq", f"{tag}_R2.fastq")
+                        else:
+                            tee.write(f"Warning: cutadapt did not produce trimmed R2 for {tag}\n")
 
                     run_cmd(
                         f"bismark -p {thread} -N 1 . -1 {tag}_R1.fastq -2 {tag}_R2.fastq")
                     run_cmd(
                         f"deduplicate_bismark -p --bam {tag}_R1_bismark_bt2_pe.bam")
-                    os.rename(f"{tag}_R1_bismark_bt2_pe.deduplicated.bam", f"{tag}.bam")
+                    if os.path.exists(f"{tag}_R1_bismark_bt2_pe.deduplicated.bam"):
+                        os.rename(f"{tag}_R1_bismark_bt2_pe.deduplicated.bam", f"{tag}.bam")
+                    else:
+                        tee.write(f"Warning: deduplicate did not produce output for {tag}\n")
                     run_cmd(
                         f"bismark_methylation_extractor --parallel {thread} -p --bedGraph "
                         f"--cutoff 4 --cytosine_report --CX --genome_folder . {tag}.bam")
@@ -109,14 +130,23 @@ def run(opts):
                         f"cutadapt -j {thread} -m 20 --trim-n -a {adaptor} -A {adaptor} "
                         f"-o {tag}_R1_trimmed.fastq -p {tag}_R2_trimmed.fastq "
                         f"{tag}_R1.fastq {tag}_R2.fastq")
-                    os.rename(f"{tag}_R1_trimmed.fastq", f"{tag}_R1.fastq")
-                    os.rename(f"{tag}_R2_trimmed.fastq", f"{tag}_R2.fastq")
+                    if os.path.exists(f"{tag}_R1_trimmed.fastq"):
+                        os.rename(f"{tag}_R1_trimmed.fastq", f"{tag}_R1.fastq")
+                    else:
+                        tee.write(f"Warning: cutadapt did not produce trimmed R1 for {tag}\n")
+                    if os.path.exists(f"{tag}_R2_trimmed.fastq"):
+                        os.rename(f"{tag}_R2_trimmed.fastq", f"{tag}_R2.fastq")
+                    else:
+                        tee.write(f"Warning: cutadapt did not produce trimmed R2 for {tag}\n")
 
                 run_cmd(
                     f"bismark -p {thread} -N 1 . -1 {tag}_R1.fastq -2 {tag}_R2.fastq")
                 run_cmd(
                     f"deduplicate_bismark -p --bam {tag}_R1_bismark_bt2_pe.bam")
-                os.rename(f"{tag}_R1_bismark_bt2_pe.deduplicated.bam", f"{tag}.bam")
+                if os.path.exists(f"{tag}_R1_bismark_bt2_pe.deduplicated.bam"):
+                    os.rename(f"{tag}_R1_bismark_bt2_pe.deduplicated.bam", f"{tag}.bam")
+                else:
+                    tee.write(f"Warning: deduplicate did not produce output for {tag}\n")
                 run_cmd(
                     f"bismark_methylation_extractor --parallel {thread} -p --bedGraph "
                     f"--cutoff 4 --cytosine_report --CX --genome_folder . {tag}.bam")
@@ -124,22 +154,22 @@ def run(opts):
                     if os.path.exists(fname):
                         os.unlink(fname)
 
-                # Generate bedgraph
-                run_cmd(
-                    f"awk '{{OFS=\"\\t\";if($4+$5>0){{"
-                    f"if($6==\"CG\"){{print $1,$2,$2+1,$4/($4+$5) > \"{tag}.CG.bedgraph\"}}; "
-                    f"if($6==\"CHG\"){{print $1,$2,$2+1,$4/($4+$5) > \"{tag}.CHG.bedgraph\"}}; "
-                    f"if($6==\"CHH\"){{print $1,$2,$2+1,$4/($4+$5) > \"{tag}.CHH.bedgraph\"}}}}}}' "
-                    f"{tag}.CX_report.txt")
-                run_cmd(
-                    f"sort -k1,1 -k2,2n {tag}.CX_report.txt > tmp")
-                run_cmd(f"bgzip -c tmp > {tag}.CX_report.txt.gz")
-                run_cmd(f"tabix -C -p vcf {tag}.CX_report.txt.gz")
-
             tee.write("\nAlignment finished...\n")
             _bin_methylation(tag, binsize, min_c, tee)
 
-            for fname in globmod.glob(f"C*_O?_*{tag}*.txt"):
+            # Generate bedgraph and compress CX_report for all branches
+            run_cmd(
+                f"awk '{{OFS=\"\\t\";if($4+$5>0){{"
+                f"if($6==\"CG\"){{print $1,$2,$2+1,$4/($4+$5) > \"{tag}.CG.bedgraph\"}}; "
+                f"if($6==\"CHG\"){{print $1,$2,$2+1,$4/($4+$5) > \"{tag}.CHG.bedgraph\"}}; "
+                f"if($6==\"CHH\"){{print $1,$2,$2+1,$4/($4+$5) > \"{tag}.CHH.bedgraph\"}}}}}}' "
+                f"{tag}.CX_report.txt")
+            run_cmd(
+                f"sort -k1,1 -k2,2n {tag}.CX_report.txt > tmp")
+            run_cmd(f"bgzip -c tmp > {tag}.CX_report.txt.gz")
+            run_cmd(f"tabix -C -p vcf {tag}.CX_report.txt.gz")
+
+            for fname in globmod.glob(f"*_OB_*{tag}*.txt") + globmod.glob(f"*_OT_*{tag}*.txt"):
                 os.unlink(fname)
             if os.path.exists("tmp"):
                 os.unlink("tmp")
@@ -149,14 +179,17 @@ def run(opts):
             run_cmd(
                 f"Rscript --vanilla {prefix}/scripts/DMRcaller.R {thread} {par_str}")
 
-        run_cmd("rm -rf Bisulfite_Genome", shell=True)
+        shutil.rmtree("Bisulfite_Genome", ignore_errors=True)
         for fname in ([f"{genome}.fasta"] + globmod.glob("*.ebwt")):
             if os.path.exists(fname):
                 os.unlink(fname)
 
     else:
         for pre in tags:
-            os.symlink(f"../{pre}.CX_report.txt.gz", f"{pre}.CX_report.txt.gz")
+            symlink_target = f"{pre}.CX_report.txt.gz"
+            if os.path.islink(symlink_target) or os.path.exists(symlink_target):
+                os.unlink(symlink_target)
+            os.symlink(f"../{pre}.CX_report.txt.gz", symlink_target)
 
         tee.write("\nPerforming DMRcaller...\n")
         run_cmd(
@@ -180,7 +213,6 @@ def _bin_methylation(tag, binsize, min_c, tee):
     cavg = {c: 0.0 for c in contexts}
     num = {c: 0 for c in contexts}
     covtotal = {c: 0 for c in contexts}
-    sum_vals = {c: 0 for c in contexts}
     totalbin = {c: 0 for c in contexts}
     passedbin = {c: 0 for c in contexts}
 
@@ -204,7 +236,7 @@ def _bin_methylation(tag, binsize, min_c, tee):
             if len(cols) < 6:
                 continue
 
-            if line_num == 0:
+            if chr_name is None:
                 chr_name = cols[0]
 
             curr_chr = cols[0]
@@ -230,7 +262,6 @@ def _bin_methylation(tag, binsize, min_c, tee):
             ctx_type = cols[5]
             ccount = int(cols[3])
             ctcount = int(cols[4])
-            sum_vals[ctx_type] += 1
 
             while pos >= flag:
                 for cont, ctx in enumerate(contexts):
